refactor(calendar): decompose calendar_service into a package of sub-1000-line submodules behind a re-export facade with no behaviour change
